Split navigation bar to components Fixes #5
